requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.0.0
fastmcp>=0.4.1

# CLI and UI
typer>=0.9.0
rich>=10.0.0
prompt_toolkit>=3.0.0

# Tools and utilities
tiktoken>=0.3.0
tokenizers>=0.13.0
regex>=2022.0.0
GitPython>=3.1.0
pygments>=2.15.0

# Performance
tqdm>=4.65.0
concurrent-log-handler>=0.9.0

# Machine Learning and Optimization
torch>=2.0.0
numpy>=1.20.0
sentence-transformers>=2.2.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Web API
fastapi>=0.100.0
uvicorn>=0.23.0
